@@ -4,7 +4,7 @@ project(indexer_connector)
 
 enable_testing()
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 add_definitions(-DPROMISE_TYPE=PromiseType::NORMAL)
