@@ -10,14 +10,15 @@
  */
 
 #include "serverSelector_test.hpp"
-#include "serverSelector.hpp"
-#include <chrono>
+#include "monitoring.hpp"
 #include <memory>
+#include <stdexcept>
 #include <string>
-#include <thread>
 
-// Healt check interval for the servers
-constexpr auto SERVER_SELECTOR_HEALTH_CHECK_INTERVAL {5u};
+// Health check interval for the servers
+constexpr auto SERVER_SELECTOR_HEALTH_CHECK_INTERVAL = 1u;
+constexpr auto SERVER_SELECTOR_HEALTH_CHECK_INTERVAL_ZERO = 0u;
+constexpr auto SERVER_SELECTOR_HEALTH_CHECK_INTERVAL_INFINITE = std::numeric_limits<uint32_t>::max();
 
 namespace Log
 {
@@ -28,95 +29,230 @@ namespace Log
 
 /**
  * @brief Test instantiation with valid servers.
- *
  */
 TEST_F(ServerSelectorTest, TestInstantiation)
 {
-    EXPECT_NO_THROW(m_selector = std::make_shared<ServerSelector>(m_servers, SERVER_SELECTOR_HEALTH_CHECK_INTERVAL));
+    std::vector<std::string> servers;
+    servers.emplace_back("http://localhost:9209");
+    servers.emplace_back("http://localhost:9210");
+    servers.emplace_back("http://localhost:9211");
+
+    // Setup mock to simulate health check responses (green and yellow servers available)
+    setupHealthCheckMocks("green", "red", "yellow");
+
+    // Test that creating the server selector doesn't throw
+    std::shared_ptr<TestServerSelector> selector;
+    EXPECT_NO_THROW(
+        selector = std::make_shared<TestServerSelector>(
+            servers, SERVER_SELECTOR_HEALTH_CHECK_INTERVAL, SecureCommunication {}, m_mockHttpRequest.get()));
 }
 
 /**
  * @brief Test instantiation without servers.
- *
  */
 TEST_F(ServerSelectorTest, TestInstantiationWithoutServers)
 {
-    m_servers.clear();
+    // Empty servers list (simulating empty server configuration)
+    std::vector<std::string> servers;
 
-    // It doesn't throw an exception because the class ServerSelector accepts vector without servers
-    EXPECT_NO_THROW(m_selector = std::make_shared<ServerSelector>(m_servers, SERVER_SELECTOR_HEALTH_CHECK_INTERVAL));
+    // Even with empty servers, selector creation should not throw
+    std::shared_ptr<TestServerSelector> selector;
+    EXPECT_NO_THROW(
+        selector = std::make_shared<TestServerSelector>(
+            servers, SERVER_SELECTOR_HEALTH_CHECK_INTERVAL, SecureCommunication {}, m_mockHttpRequest.get()));
+
+    EXPECT_THROW(selector->getNext(), std::runtime_error);
 }
 
 /**
- * @brief Test instantiation and getNext server before health check.
- *
+ * @brief Test getNext server before health check.
  */
 TEST_F(ServerSelectorTest, TestGetNextBeforeHealthCheck)
 {
-    const auto hostGreenServer {m_servers.at(0)};
-    const auto hostYellowServer {m_servers.at(2)};
+    std::vector<std::string> servers;
+    servers.emplace_back("http://localhost:9209");
+    servers.emplace_back("http://localhost:9210");
+    servers.emplace_back("http://localhost:9211");
+
+    const auto server1 = servers.at(0);
+    const auto server3 = servers.at(2);
+
+    // Setup mock to simulate green and yellow servers available, red server down
+    setupHealthCheckMocks("green", "red", "yellow");
+
+    // Create selector
+    std::shared_ptr<TestServerSelector> selector;
+    EXPECT_NO_THROW(
+        selector = std::make_shared<TestServerSelector>(
+            servers, SERVER_SELECTOR_HEALTH_CHECK_INTERVAL_INFINITE, SecureCommunication {}, m_mockHttpRequest.get()));
 
     std::string nextServer;
 
-    EXPECT_NO_THROW(m_selector = std::make_shared<ServerSelector>(m_servers, SERVER_SELECTOR_HEALTH_CHECK_INTERVAL));
+    // Test first call returns green server (available)
+    EXPECT_NO_THROW(nextServer = selector->getNext());
+    EXPECT_EQ(nextServer, server1);
 
-    // It doesn't throw an exception because the green and yellow servers are available
-    EXPECT_NO_THROW(nextServer = m_selector->getNext());
-    EXPECT_EQ(nextServer, hostGreenServer);
-
-    // It doesn't throw an exception because the green and yellow servers are available
-    EXPECT_NO_THROW(nextServer = m_selector->getNext());
-    EXPECT_EQ(nextServer, hostYellowServer);
+    // Test second call returns yellow server (available, red is skipped)
+    EXPECT_NO_THROW(nextServer = selector->getNext());
+    EXPECT_EQ(nextServer, server3);
 }
 
 /**
- * @brief Test instantiation and getNext server before and after health check.
- *
+ * @brief Test getNext server before and after health check completion.
  */
 TEST_F(ServerSelectorTest, TestGetNextBeforeAndAfterHealthCheck)
 {
-    const auto hostGreenServer {m_servers.at(0)};
-    const auto hostYellowServer {m_servers.at(2)};
+    std::vector<std::string> servers;
+    servers.emplace_back("http://localhost:9209");
+    servers.emplace_back("http://localhost:9210");
+    servers.emplace_back("http://localhost:9211");
+
+    const auto server1 = servers.at(0);
+    const auto server3 = servers.at(2);
+
+    std::atomic<int> phase {1};
+    std::atomic<int> callCount {0};
+
+    EXPECT_CALL(*m_mockHttpRequest, get(testing::_, testing::_, testing::_))
+        .WillRepeatedly(testing::Invoke(
+            [&phase, &callCount](auto requestParams, const auto& postParams, auto /*configParams*/)
+            {
+                // Extract the URL to determine which server is being checked
+                std::string url;
+                if (std::holds_alternative<TRequestParameters<std::string>>(requestParams))
+                {
+                    url = std::get<TRequestParameters<std::string>>(requestParams).url.url();
+                }
+                else
+                {
+                    url = std::get<TRequestParameters<nlohmann::json>>(requestParams).url.url();
+                }
+
+                std::string response;
+
+                if (int currentPhase = phase.load(); currentPhase == 1)
+                {
+                    // Phase 1: green and yellow servers available, red server down
+                    if (url.find("9209") != std::string::npos)
+                    {
+                        response = R"([{"status":"green"}])";
+                    }
+                    else if (url.find("9210") != std::string::npos)
+                    {
+                        response = R"([{"status":"red"}])";
+                    }
+                    else if (url.find("9211") != std::string::npos)
+                    {
+                        response = R"([{"status":"yellow"}])";
+                    }
+                    callCount++;
+                }
+                else if (currentPhase == 2)
+                {
+                    // Phase 2: all servers down
+                    if (url.find("9209") != std::string::npos || url.find("9210") != std::string::npos ||
+                        url.find("9211") != std::string::npos)
+                    {
+                        response = R"([{"status":"red"}])";
+                    }
+                    callCount++;
+                }
+                else if (currentPhase == 3)
+                {
+                    // Phase 3: only server 1 available
+                    if (url.find("9209") != std::string::npos)
+                    {
+                        response = R"([{"status":"green"}])";
+                    }
+                    else if (url.find("9210") != std::string::npos || url.find("9211") != std::string::npos)
+                    {
+                        response = R"([{"status":"red"}])";
+                    }
+                    callCount++;
+                }
+
+                if (!response.empty())
+                {
+                    postParams.onSuccess(response);
+                }
+                else
+                {
+                    postParams.onError("Server not found", 404);
+                }
+            }));
+
+    // Create selector with a short interval for quick health checks
+    std::shared_ptr<TestServerSelector> selector;
+    EXPECT_NO_THROW(
+        selector = std::make_shared<TestServerSelector>(
+            servers, SERVER_SELECTOR_HEALTH_CHECK_INTERVAL_ZERO, SecureCommunication {}, m_mockHttpRequest.get()));
 
     std::string nextServer;
 
-    EXPECT_NO_THROW(m_selector = std::make_shared<ServerSelector>(m_servers, SERVER_SELECTOR_HEALTH_CHECK_INTERVAL));
+    while (callCount < 2)
+    {
+        std::this_thread::yield();
+    }
 
-    // It doesn't throw an exception because yellow and green servers are available before health check
-    EXPECT_NO_THROW(nextServer = m_selector->getNext());
-    EXPECT_EQ(nextServer, hostGreenServer);
+    // Before health check - green and yellow servers are available
+    EXPECT_NO_THROW(nextServer = selector->getNext());
+    EXPECT_EQ(nextServer, server1);
 
-    // It doesn't throw an exception because the green and yellow servers are available
-    EXPECT_NO_THROW(nextServer = m_selector->getNext());
-    EXPECT_EQ(nextServer, hostYellowServer);
+    EXPECT_NO_THROW(nextServer = selector->getNext());
+    EXPECT_EQ(nextServer, server3);
 
-    // Interval to check the health of the servers
-    std::this_thread::sleep_for(std::chrono::seconds(SERVER_SELECTOR_HEALTH_CHECK_INTERVAL + 5));
+    EXPECT_NO_THROW(nextServer = selector->getNext());
+    EXPECT_EQ(nextServer, server1);
 
-    // Next server will be the green because is the next available server
-    EXPECT_NO_THROW(nextServer = m_selector->getNext());
-    EXPECT_EQ(nextServer, hostGreenServer);
+    EXPECT_NO_THROW(nextServer = selector->getNext());
+    EXPECT_EQ(nextServer, server3);
 
-    // Next server will be the yellow because the red server isn't available
-    EXPECT_NO_THROW(nextServer = m_selector->getNext());
-    EXPECT_EQ(nextServer, hostYellowServer);
+    phase = 2;
+    callCount = 0;
+
+    while (callCount < 2)
+    {
+        std::this_thread::yield();
+    }
+
+    // After health check all servers are down
+    EXPECT_THROW(nextServer = selector->getNext(), std::runtime_error);
+
+    phase = 3;
+    callCount = 0;
+    while (callCount < 2)
+    {
+        std::this_thread::yield();
+    }
+
+    // After health check server 1 is available
+    EXPECT_NO_THROW(nextServer = selector->getNext());
+    EXPECT_EQ(nextServer, server1);
+
+    EXPECT_NO_THROW(nextServer = selector->getNext());
+    EXPECT_EQ(nextServer, server1);
 }
 
 /**
- * @brief Test instantiation and getNext when there are no available servers.
- *
+ * @brief Test getNext when there are no available servers.
  */
-TEST_F(ServerSelectorTest, TestGextNextWhenThereAreNoAvailableServers)
+TEST_F(ServerSelectorTest, TestGetNextWhenThereAreNoAvailableServers)
 {
-    const auto hostRedServer {m_servers.at(1)};
+    const auto hostRedServer = "http://localhost:9210";
 
-    m_servers.clear();
-    m_servers.emplace_back(hostRedServer);
+    // Create servers with only red server (which will be down)
+    std::vector<std::string> servers;
+    servers.emplace_back(hostRedServer);
 
-    std::string nextServer;
+    // Setup mock to simulate only red server (down)
+    setupHealthCheckMocks("red", "red", "red");
 
-    EXPECT_NO_THROW(m_selector = std::make_shared<ServerSelector>(m_servers, SERVER_SELECTOR_HEALTH_CHECK_INTERVAL));
+    // Create selector with only the red server
+    std::shared_ptr<TestServerSelector> selector;
+    EXPECT_NO_THROW(
+        selector = std::make_shared<TestServerSelector>(
+            servers, SERVER_SELECTOR_HEALTH_CHECK_INTERVAL, SecureCommunication {}, m_mockHttpRequest.get()));
 
-    // It throws an exception because there are no available servers
-    EXPECT_THROW(nextServer = m_selector->getNext(), std::runtime_error);
+    // Test that exception is thrown when no servers are available
+    EXPECT_THROW(selector->getNext(), std::runtime_error);
 }
